requirements.txt
flask==3.0.3
flask-wtf==1.2.1
flask-login==0.6.3
argon2-cffi==23.1.0
bleach==6.1.0
python-dotenv==1.0.1
Werkzeug==3.0.3
gunicorn==21.2.0
